8382616: Create GCArguments::set_heap_size interface
